add: some small changes

+++ onix/modules/soundCouldDl.php
@@ -14,20 +14,23 @@ if ($text == '「 📻 دانلود ساندکلود 」'){
     $soundInfo = $apiRequest->getSoundCloudInfo($response)->result;
     $soundName = str_replace('.mp3', '', $soundInfo->title);
 
-    $bot->editMessage($chat_id, '<b>فایل مورد نظر یافت شد! درحال ارسال به تلگرام ...</b>', $message_id + 1);
-    $bot->sendChatAction($from_id, 'upload_document');
-
-    $getFile = $apiRequest->getSoundCloudFile($response);
-
-    $rand = rand(124588, 854963);
-    $audioFile = "SoundCloud-{$response}.mp3";
-
-    file_put_contents($audioFile, $getFile);
-    $bot->sendAudio($from_id, "https://fara-it.ir/onix/{$audioFile}", $soundName, $downloaderKeyboard);
-    $userCursor->setLimit($from_id, 'dl_soundcloud', $userLimits->dl_soundcloud - 1);
-    unlink($audioFile);
-    $userCursor->setLimit($from_id , 'dl_soundcloud' , $userLimits->dl_soundcloud - 1);
-
+    if($response){
+        $bot->editMessage($chat_id, '<b>فایل مورد نظر یافت شد! درحال ارسال به تلگرام ...</b>', $message_id + 1);
+        $bot->sendChatAction($from_id, 'upload_document');
+        
+        $getFile = $apiRequest->getSoundCloudFile($response);
+    
+        $rand = rand(124588, 854963);
+        $audioFile = "SoundCloud-{$response}.mp3";
+    
+        file_put_contents($audioFile, $getFile);
+        $bot->sendAudio($from_id, "https://fara-it.ir/onix/{$audioFile}", $soundName, $downloaderKeyboard);
+        unlink($audioFile);
+        $userCursor->setLimit($from_id , 'dl_soundcloud' , $userLimits->dl_soundcloud - 1);
+        $userCursor->setStep($from_id, 'home');
+    }else{
+        $bot->editMessage($chat_id, '<b>فایل مورد نظر یافت نشد!  ...</b>', $message_id + 1);
+    }
 
     die;
 }
